requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
APScheduler==3.11.0
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
binaryornot==0.4.4
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
cryptography==45.0.4
cssselect2==0.8.0
Django==5.2.3
django-browser-reload==1.18.0
django-tailwind==4.0.1
frozenlist==1.7.0
html5lib==1.1
idna==3.10
Jinja2==3.1.6
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
netifaces==0.11.0
oscrypto==1.3.0
pillow==11.1.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
Pygments==2.19.2
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pypdf==5.6.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-nmap==0.7.1
python-slugify==8.0.4
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.4
rich==14.0.0
scapy==2.6.1
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.4.0
twilio==9.6.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17
yarl==1.20.1
